feat(rpc): implement Filecoin.StateSectorPreCommitInfo
